# improve prng compile times with loop rolling