--- a/jax/random.py
+++ b/jax/random.py
@@ -123,33 +123,19 @@ def threefry_2x32(keypair, count):
   else:
     x = list(np.split(count.ravel(), 2))
 
-  rotations = asarray([13, 15, 26, 6, 17, 29, 16, 24], dtype="uint32")
-  ks = [key1, key2, key1 ^ key2 ^ onp.uint32(0x1BD11BDA)]
+  rotations = asarray([[13, 15, 26, 6], [17, 29, 16, 24]], dtype="uint32")
+  ks = asarray([key1, key2, key1 ^ key2 ^ onp.uint32(0x1BD11BDA)], dtype="uint32")
+  idxs = asarray([[1, 2], [2, 0], [0, 1], [1, 2], [2, 0]])
 
   x[0] = x[0] + ks[0]
   x[1] = x[1] + ks[1]
 
-  x = lax.fori_loop(0, 4, lambda i, x: apply_round(x, rotations[i]), x)
-  x[0] = x[0] + ks[1]
-  x[1] = x[1] + ks[2] + onp.uint32(1)
-
-  x = lax.fori_loop(4, 8, lambda i, x: apply_round(x, rotations[i]), x)
-  x[0] = x[0] + ks[2]
-  x[1] = x[1] + ks[0] + onp.uint32(2)
-
-  x = lax.fori_loop(0, 4, lambda i, x: apply_round(x, rotations[i]), x)
-  x[0] = x[0] + ks[0]
-  x[1] = x[1] + ks[1] + onp.uint32(3)
-
-  x = lax.fori_loop(4, 8, lambda i, x: apply_round(x, rotations[i]), x)
-  x[0] = x[0] + ks[1]
-  x[1] = x[1] + ks[2] + onp.uint32(4)
-
-  x = lax.fori_loop(0, 4, lambda i, x: apply_round(x, rotations[i]), x)
-  x[0] = x[0] + ks[2]
-  x[1] = x[1] + ks[0] + onp.uint32(5)
-
-  out = np.concatenate(x)
+  def step(i, x):
+    for r in rotations[i % 2]:
+      x = apply_round(x, r)
+    i0, i1 = idxs[i]
+    return [x[0] + ks[i0], x[1] + ks[i1] + asarray(i + 1, dtype="uint32")]
+  out = np.concatenate(lax.fori_loop(0, 5, step, x))
   assert out.dtype == onp.uint32
   return lax.reshape(out[:-1] if odd_size else out, count.shape)