@fehiepsi 2019-08-13
Thanks so much @mattjj ! This works pretty well! The only side effect which I observe is the trick of using `random.split(key, 1)` in gamma sampler does not work anymore with this PR. Anyway, it is minor and I think that it is better to avoid using that trick. I will try to explore warp-based parallelism as proposed in #552 instead.